ReconocimientoV2/requirements.txt
# Librerías principales
imageio>=2.34.0
mtcnn>=0.1.1
keras-facenet>=0.3.2
numpy>=1.24.0
Pillow>=10.0.0

# Dependencias necesarias para keras-facenet y mtcnn
tensorflow>=2.13.0
keras>=2.13.0
